perf(variants): the multi-sample cell loops read genotype codes, not a record
A per-feature Record<sampleName, genotype> was built by @gmod/vcf's GENOTYPES()
and then walked three more times — computeSampleInfo for the ploidy/phasing and
legend flags, each cell loop for the colors, internFeatureGenotypes to turn it
back into the codes the payload actually ships. Four traversals and F x S string
allocations plus a dictionary-mode object per feature, to reproduce something
the worker never sends.

computeSampleInfo now makes one processGenotypes pass instead, interning
straight off the (line, start, end) ranges @gmod/vcf reports. A per-site memo of
the ranges already seen answers almost every sample without materializing its
substring at all — a site carries a handful of distinct genotypes across
thousands of samples — and carries the classification with the code, so the char
walk also runs once per distinct genotype rather than once per cell. The cell
loops index those codes by a source's column, resolved once per pass, and key
their style memos by code, which makes them array reads where they were string
hashes. 2504 samples x 400 variants: 611ms/214MB -> 71ms/1MB.

Codes widen to Uint32. The Uint16 cap of 65535 distinct genotype strings was
reachable on a decomposed pangenome callset, and past it a genotype interned to
0 — survivable when the cells were painted from the strings and only hover lost
the call, but 0 now means "no call" to the loops that color from the codes, so
the overflow would simply not be painted.

Transferables become a Set: one codes array is now shared by every reference to
its feature rather than rebuilt per shipped entry, and getFeaturesInMultipleRegions
merges without deduping, so a variant spanning two displayed regions would
otherwise offer the same buffer to postMessage twice.

+++ plugins/variants/src/shared/MultiSampleVariantBaseModel.ts
@@ -113,7 +113,7 @@ const PORTABLE_CONFIG_KEYS = [
 // every fetched region, while the codes live per-region in regular mode and in
 // one flat array in matrix mode.
 function getOrderedGenotypeCodes(cellData: CellDataResult) {
-  const genotypeCodesByFeatureId = new Map<string, Uint16Array>()
+  const genotypeCodesByFeatureId = new Map<string, Uint32Array>()
   if (cellData.mode === 'regular') {
     for (const regionData of Object.values(cellData.perRegionCellData)) {
       for (const featureId in regionData.featureGenotypeMap) {
